fix slugification logic
The previous logic broke upon facing a file name with a `'` in it (for instance, `The Attack on the Senses by Eddington's Two Tables.md`). This fixes it by filtering out any non-slug friendly characters at the end of the pipeline. Also now needs to slugify the file comparison to determine if the note exists or not.