use positive lookbehind on regex
Previously the regex would match on every leading `[` of a wikilink, so if I have "Ayn Rand offers a new formulation of \[[[The Law of Identity|the law of identity]]\]:" it would remove the escaped `[` character. This lookbehind fixes this. Browser compatibility should not be an issue, because all of that bs is processed by node, so only one regex engine to care about.